feat: guest 방명록 모달 세부사항 조정
