+++ frontend/src/components/specific/modal/guestPhotoModal/GuestPhotoModal.styles.ts
@@ -2,7 +2,6 @@ import styled from '@emotion/styled';
 
 export const Wrapper = styled.div`
 width: 100%;
-z-index: ${({ theme }) => theme.zIndex.modal};
 display: flex;
     flex-direction: column;
     gap: 8px;
