+++ frontend/src/pages/guest/guestbookPage/card/GuestGuestbookCardPage.styles.ts
@@ -17,7 +17,6 @@ export const InfoSection = styled.section`
 
 export const DeleteButtonContainer = styled.div`
   display: flex;
-  justify-content: space-between;
 `;
 
 export const InfoTitle = styled.h1`
